// src/MVP.Project.UI.Web/src/app/models/customer.view-model.ts
export interface CustomerViewModel {
    id?: string;
    name: string;
    email: string;
    documentNumber: string;
    birthDate: string;
    phone?: string;
    stateInscription?: string;
    streetAddress?: string;
    buildingNumber?: string;
    secondaryAddress?: string;
    neighborhood?: string;
    zipCode?: string;
    city?: string;
    state?: string;
    active: boolean;
} 